syscalls: remove ARM7-specific syscalls.c file
